requirements.txt
python-dotenv
python-telegram-bot==20.3
selenium
webdriver-manager
beautifulsoup4
requests
pymongo